🔨 chore: split `genUserLLMConfig`
