+++ src/const/settings/llm.ts
@@ -1,6 +1,6 @@
 import { ModelProvider } from '@/libs/agent-runtime';
 
-import { genUserLLMConfig } from '@/server/globalConfig/genLLMConfig'
+import { genUserLLMConfig } from '@/utils/genUserLLMConfig'
 
 export const DEFAULT_LLM_CONFIG = genUserLLMConfig({
   ollama: {
