+++ src/server/globalConfig/genServerLLMConfig.ts
@@ -6,7 +6,6 @@ import { ModelProvider } from '@/libs/agent-runtime';
 import { extractEnabledModels, transformToChatModelCards } from '@/utils/parseModels';
 
 import { ModelProviderCard } from '@/types/llm';
-import { UserModelProviderConfig } from '@/types/user/settings';
 
 export const genServerLLMConfig = () => {
   const llmConfig = getLLMConfig() as Record<string, any>;
@@ -47,20 +46,3 @@ export const genServerLLMConfig = () => {
     return config;
   }, {} as Record<ModelProvider, any>);
 };
-
-export const genUserLLMConfig = (specificConfig: Record<any, any>): UserModelProviderConfig => {
-  return Object.keys(ModelProvider).reduce((config, providerKey) => {
-    const provider = ModelProvider[providerKey as keyof typeof ModelProvider];
-    const providerCard = ProviderCards[`${providerKey}ProviderCard` as keyof typeof ProviderCards] as ModelProviderCard;
-    const providerConfig = specificConfig[provider as keyof typeof specificConfig] || {};
-
-    config[provider] = {
-      enabled: providerConfig.enabled !== undefined ? providerConfig.enabled : false,
-      enabledModels: providerCard ? ProviderCards.filterEnabledModels(providerCard) : [],
-      ...(providerConfig.fetchOnClient !== undefined && { fetchOnClient: providerConfig.fetchOnClient }),
-    };
-
-    return config;
-  }, {} as UserModelProviderConfig);
-};
-
